+++ src/Restaurants/Restaurants.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
 import Header from "./Header/Header";
-import Footer from "../components/Footer/Footer";
+import Footer from "./Footer/Footer";
 
 const Restaurants = () => {
   return (
